Use Erlang's and Elixir's more complete log levels
Starting with Elixir 1.11, Elixir started using the full list of log
levels provided by Erlang's new logger service. This commit the
following:

1. Remove code to convert syslog severity to Elixir log levels
2. Rename `:informational` to `:info` to match Erlang and Elixir
3. Remove the `:severity` field from the log message since it's
   redundant now.

This is technically a breaking change for any code that used the
`:severify` field in the log message, but I don't think there are any
cases in the wild that do that. I'd expect most code to use the log
level and after this commit, it will match the reported level.