Convert to a library
This is a backwards incompatible change.

Instead of starting up the kmsg and syslog tailers automatically. They
now must be linked into a supervision tree. This means that
NervesLogging no longer needs to figure out whether it's running on a
device or not. That simplifies the code here and makes it possible for
the code that uses it to decide.