Fix logging so it was actually adjustable
There are a few ways of adjusting logging levels with the Erlang logger,
but the ones in the README.md actually didn't work. The problems were that
the existing code was calling `Logger.bare_log` and not setting the
`:mfa` metadata to make the caller identifiable AND `Logger.bare_log`
doesn't do any checking of `:mfa` so filtering at the Logger call
doesn't work.

`Logger.log/3` makes this easy since it fills in `:mfa` automatically
and filters on the calling module before forwarding the message.